refine bufferPool
