+++ lib/core/base-renderer.js
@@ -6,6 +6,7 @@ import shaderChunks from '../shaders/chunks/index';
 import shaderTemplates from '../shaders/templates/index';
 import ProgramLib from './program-lib';
 import View from './view';
+import BufferPool from './buffer-pool';
 
 let _m3_tmp = mat3.create();
 let _m4_tmp = mat4.create();
@@ -312,6 +313,8 @@ export default class BaseRenderer {
     this._stage2fn = {};
     this._modelType2fn = {};
     this._usedTextureUnits = 0;
+
+    this._bufferPool = new BufferPool();
 
     this._viewPools = new RecyclePool(() => {
       return new View();
