+++ lib/renderers/forward-renderer.js
@@ -1,7 +1,6 @@
 import gfx from 'gfx.js';
 import { vec3, mat4 } from 'vmath';
 import { CircularPool } from 'memop';
-import BufferPool from './buffer-pool';
 
 import BaseRenderer from '../core/base-renderer';
 import InputAssembler from '../core/input-assembler';
@@ -19,8 +18,7 @@ let _a16_viewProj = new Float32Array(16);
 let _a16_lightViewProj = new Float32Array(16);
 let _a3_camPos = new Float32Array(3);
 
-let _bufferPool = new BufferPool();
-const MAX_LINE_VERTS = Math.ceil(_bufferPool.maxBytes / 24); // pos: 3, color: 3
+let _maxLineVerts = -1;
 
 export default class ForwardRenderer extends BaseRenderer {
   constructor(device, builtin) {
@@ -41,6 +39,7 @@ export default class ForwardRenderer extends BaseRenderer {
     this._registerModel('skinning', this._drawSkinning.bind(this));
 
     // this._blur = new GaussianBlur(device, this._programLib);
+    _maxLineVerts = Math.ceil(this._bufferPool.maxBytes / 24); // pos: 3, color: 3
 
     this._lineIAs = new CircularPool(() => {
       return new InputAssembler(
@@ -52,7 +51,7 @@ export default class ForwardRenderer extends BaseRenderer {
           ]),
           gfx.USAGE_DYNAMIC,
           null,
-          MAX_LINE_VERTS
+          _maxLineVerts
         ),
         null,
         gfx.PT_LINES
@@ -358,12 +357,12 @@ export default class ForwardRenderer extends BaseRenderer {
     let model = item.model;
     let v = 0;
     // model.line total: 48 bytes (start: vec3, end: vec3, color: vec3, normal: vec3)
-    let lineData = _bufferPool.request(model.lines.length * 48);
+    let lineData = this._bufferPool.request(model.lines.length * 48);
     let float32LineData = lineData.float32View;
 
     for (let i = model.lines.length-1; i >= 0; --i) {
       // flush when verts exceeds
-      if (v + 2 >= MAX_LINE_VERTS) {
+      if (v + 2 >= _maxLineVerts) {
         let lineIA = this._lineIAs.request();
         lineIA._vertexBuffer.update(0, float32LineData);
         lineIA._start = 0;
